requirements.txt
langchain
faiss
langchain_community
langchain_openai
python-dotenv
streamlit
HuggingFace
sentence_transformers
